Initial upload of project files and rubric
This commit includes the initial set of project files for the Data 601 course project on Uber pickups in NYC. The Jupyter notebook contains detailed data analysis and visualization codes. The PowerPoint file provides the project rubric and expectations. Also included is a ZIP archive with additional data and resources necessary for the project.